chore: example update
