+++ _examples/scrapejsp/scrapejsp/constants.go
@@ -1,3 +1,3 @@
-package scrapeThisSite
+package scrapejsp
 
 // you can define your constants here
