+++ _examples/scrapejsp/scrapejsp/errors.go
@@ -1,0 +1,3 @@
+package scrapejsp
+
+// you can define your errors here
