+++ _examples/scrapejsp/utils/helper.go
