--- _examples/scrapethissite/scrapethissite/core.go
@@ -1,49 +1,0 @@
-package scrapeThisSite
-
-// do not modify this file
-
-import (
-	"context"
-
-	"github.com/tech-engine/goscrapy/pkg/core"
-)
-
-type coreSpider struct {
-	ctx       context.Context
-	delegator core.Delegator[*Job, []Record]
-}
-
-func (cs *coreSpider) NewJob(id string) *Job {
-	return NewJob(id)
-}
-
-func (cs *coreSpider) SetDelegator(delegator core.Delegator[*Job, []Record]) {
-	cs.delegator = delegator
-}
-
-func (cs *coreSpider) NewRequest() *core.GSRequest {
-	return cs.delegator.NewRequest()
-}
-
-func (cs *coreSpider) Request(ctx context.Context, req *core.GSRequest, cb core.ResponseCallback) {
-	cs.delegator.ExRequest(ctx, req, cb)
-}
-
-func (cs *coreSpider) yield(output *Output) {
-	cs.delegator.Yield(output)
-}
-
-func (cs *coreSpider) jobFromContext(ctx context.Context) (*Job, error) {
-	meta, ok := ctx.Value("META_DATA").(map[string]any)
-
-	if !ok {
-		return nil, ERR_EXTRACTING_META
-	}
-
-	job, ok := meta["JOB"].(*Job)
-
-	if !ok {
-		return nil, ERR_EXTRACTING_JOB
-	}
-	return job, nil
-}
